+++ ReleaseNotes.md
@@ -1,3 +1,9 @@
+# Version: 0.0.13dev1
+Date: 2024.12.22
+- Changes:
+  - Download Cover from MusicBrainz
+  - Fix selection of whisper model
+
 # Version: 0.0.12
 Date: 2024.12.19
 - Changes:
@@ -12,6 +18,7 @@ Date: 2024.12.19
   - Fix model path option
   - Fix error in PDF sheet generation
   - Fix hypen language download
+  - Fix install scripts
   - Some bug fixes and improved error handling and logs
 
 # Version: 0.0.11